rename distributed_env.py to env.py
